pyproject.toml
[project]
name = "ergosphere-solution"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "django>=4.2,<5.0",
    "djangorestframework>=3.14",
    "django-cors-headers>=4.3",
    "django-environ>=0.11",
    "drf-spectacular>=0.27",
    "mysqlclient>=2.2",
    "selenium>=4.15",
    "beautifulsoup4>=4.12",
    "webdriver-manager>=4.0",
    "chromadb>=0.4",
    "sentence-transformers>=2.2",
    "groq>=0.4",
]
